+++ ZXingObjC/oned/ZXEAN13Writer.m
@@ -31,7 +31,7 @@ const int ZX_EAN13_CODE_WIDTH = 3 + // start guard
 - (ZXBitMatrix *)encode:(NSString *)contents format:(ZXBarcodeFormat)format width:(int)width height:(int)height hints:(ZXEncodeHints *)hints error:(NSError **)error {
   if (format != kBarcodeFormatEan13) {
     @throw [NSException exceptionWithName:NSInvalidArgumentException
-                                   reason:[NSString stringWithFormat:@"Can only encode EAN_13, but got %d", format]
+                                   reason:[NSString stringWithFormat:@"Can only encode EAN_13, but got %lu", (unsigned long)format]
                                  userInfo:nil];
   }
